fix mocking tests
